Fix for ajax posting not showing you your new post
- TODO: there are console.logs that still need to be removed
- Posting was loading the new page twice - once in ajax.js and once in
auto-reload.js. auto-reload will handle this behaviour, the other is
commented out
- The new post does not immediately show up when immediately ajax
requesting the page after posting completes. Adding an epoch query
parameter did not change this. Instead of figuring out why it was easier
to add a 500ms delay before requesting the page, now it seems to work
well.