Upgrade Twig library for PHP 7.4 compatibility
The existing Twig library was using outdated PHP constructs. This
commit adds the foundation for using composer to pull the Twig
dependency. We make some minor changes to how Twig is initialized.

At some point in the past, the `trans` tag was removed from the Twig
core library. I eventually found the twig/extensions library which
implements the `trans` tag. In newer versions of Twig, the I18n
support seems to be implemented in the Symfony Bridge library. When
the time comes to upgrade to Twig 3.x, we will need to take this into
consideration.

You will notice that we are still dependent on the old Twig
library. This is because the Tinyboard extension still needs to be
rewritten. We will have to address that in the future.