Allow posting from catalog.
Addresses #134.

The catalog template has been modified to use the board array, which
is expected by the post form template. Because the overboard is a
"virtual board", we created a "fake" overboard array to supply
necessary information to the modified template.

Javascript was also added to hide form on first visit.