More fixes after upgrading to php8 broke everything
- call openBoard inside of catalog build. I have no idea
how it could have been defined before because it depends on a global
variable $board which seems to be only globally defined by openBoard in
functions.php, but that's called later. the fix is to just call
openBoard, that's what tinyboard does.

(somehow the code expected the uri to already be set)